Change output format

+++ src/ck-perf/trace-summary.h
@@ -34,29 +34,29 @@ class BinEntry {
                  _msgSize(0), _msgCount(0),
                  _recvSize(0), _recvCount(0)
     {
-      _msgCountPerEP = CkVec<int>(_entryTable.size() + 10);
-      _msgSizePerEP = CkVec<int>(_entryTable.size() + 10);
-      _recvCountPerEP = CkVec<int>(_entryTable.size() + 10);
-      _recvSizePerEP = CkVec<int>(_entryTable.size() + 10);
+      _msgCountPerEP = CkVec<long>(_entryTable.size() + 10);
+      _msgSizePerEP = CkVec<long>(_entryTable.size() + 10);
+      _recvCountPerEP = CkVec<long>(_entryTable.size() + 10);
+      _recvSizePerEP = CkVec<long>(_entryTable.size() + 10);
       for(int i = 0; i < _msgCountPerEP.size(); ++i) {
         _msgCountPerEP[i] = _msgSizePerEP[i] = _recvCountPerEP[i] = _recvSizePerEP[i] = 0;
       }
     }
 
-    BinEntry(double t, double idleT, int msgSize, int msgCount,
-             CkVec<int> msgSizePerEP, CkVec<int> msgCountPerEP,
-             int recvSize, int recvCount,
-             CkVec<int> recvSizePerEP, CkVec<int> recvCountPerEP
+    BinEntry(double t, double idleT, long msgSize, long msgCount,
+             CkVec<long> msgSizePerEP, CkVec<long> msgCountPerEP,
+             long recvSize, long recvCount,
+             CkVec<long> recvSizePerEP, CkVec<long> recvCountPerEP
              ): _time(t), _idleTime(idleT),
                  _msgSize(msgSize), _msgCount(msgCount),
                  _recvSize(recvSize), _recvCount(recvCount)
     {
       if(_msgSizePerEP.size() == 0)
       {
-        _msgCountPerEP = CkVec<int>(msgSizePerEP.size());
-        _msgSizePerEP = CkVec<int>(msgSizePerEP.size());
-        _recvCountPerEP = CkVec<int>(msgSizePerEP.size());
-        _recvSizePerEP = CkVec<int>(msgSizePerEP.size());
+        _msgCountPerEP = CkVec<long>(msgSizePerEP.size());
+        _msgSizePerEP = CkVec<long>(msgSizePerEP.size());
+        _recvCountPerEP = CkVec<long>(msgSizePerEP.size());
+        _recvSizePerEP = CkVec<long>(msgSizePerEP.size());
       }
       for(int i = 0; i < msgSizePerEP.size(); ++i) {
         _msgSizePerEP[i] = msgSizePerEP[i];
@@ -67,14 +67,14 @@ class BinEntry {
   }
     double &time() { return _time; }
     double &getIdleTime() { return _idleTime; }
-    int &getSize() { return _msgSize; }
-    int &getCount() {return _msgCount; }
-    CkVec<int> &getSizePerEP() { return _msgSizePerEP;}
-    CkVec<int> &getCountPerEP() {return _msgCountPerEP; }
-    int &getRecvSize() { return _recvSize; }
-    int &getRecvCount() { return _recvCount; }
-    CkVec<int> &getRecvSizePerEP() {return _recvSizePerEP; }
-    CkVec<int> &getRecvCountPerEP() {return _recvCountPerEP; }
+    long &getSize() { return _msgSize; }
+    long &getCount() {return _msgCount; }
+    CkVec<long> &getSizePerEP() { return _msgSizePerEP;}
+    CkVec<long> &getCountPerEP() {return _msgCountPerEP; }
+    long &getRecvSize() { return _recvSize; }
+    long &getRecvCount() { return _recvCount; }
+    CkVec<long> &getRecvSizePerEP() {return _recvSizePerEP; }
+    CkVec<long> &getRecvCountPerEP() {return _recvCountPerEP; }
     void write(FILE *fp);
     int  getU();
     int getUIdle();
@@ -84,12 +84,12 @@ class BinEntry {
     double _idleTime;
     int _msgSize;
     int _msgCount;
-    CkVec<int> _msgSizePerEP;
-    CkVec<int> _msgCountPerEP;
-    int _recvSize;
-    int _recvCount;
-    CkVec<int> _recvSizePerEP;
-    CkVec<int> _recvCountPerEP;
+    CkVec<long> _msgSizePerEP;
+    CkVec<long> _msgCountPerEP;
+    long _recvSize;
+    long _recvCount;
+    CkVec<long> _recvSizePerEP;
+    CkVec<long> _recvCountPerEP;
 };
 
 /// a phase entry for trace summary
@@ -242,10 +242,10 @@ class SumLogPool {
     void initMem();
     void write(void) ;
     void writeSts(void);
-    void add(double time, double idleTime, int msgSize, int msgCount,
-             CkVec<int> msgSizePerEP, CkVec<int> msgCountPerEP,
-             int recvSize, int recvCount,
-             CkVec<int> recvSizePerEP, CkVec<int> recvCountPerEP,
+    void add(double time, double idleTime, long msgSize, long msgCount,
+             CkVec<long> msgSizePerEP, CkVec<long> msgCountPerEP,
+             long recvSize, long recvCount,
+             CkVec<long> recvSizePerEP, CkVec<long> recvCountPerEP,
              int pe);
     void setEp(int epidx, double time);
     void clearEps() {
@@ -314,14 +314,14 @@ class TraceSummary : public Trace {
     int inIdle;
     int inExec;
     int depth;
-    int msgSize;
-    int msgCount;
-    CkVec<int> msgSizePerEP;
-    CkVec<int> msgCountPerEP;
-    int recvSize;
-    int recvCount;
-    CkVec<int> recvSizePerEP;
-    CkVec<int> recvCountPerEP;
+    long msgSize;
+    long msgCount;
+    CkVec<long> msgSizePerEP;
+    CkVec<long> msgCountPerEP;
+    long recvSize;
+    long recvCount;
+    CkVec<long> recvSizePerEP;
+    CkVec<long> recvCountPerEP;
 
   public:
     TraceSummary(char **argv);
